+++ samples/android-app/src/androidTest/java/org/jlleitschuh/gradle/ktlint/android/ExampleInstrumentedTest.kt
@@ -1,22 +1,19 @@
 package org.jlleitschuh.gradle.ktlint.android
 
-import android.support.test.InstrumentationRegistry
-import android.support.test.runner.AndroidJUnit4
+import androidx.test.platform.app.InstrumentationRegistry
 import org.junit.Assert.assertEquals
 import org.junit.Test
-import org.junit.runner.RunWith
 
 /**
  * Instrumented test, which will execute on an Android device.
  *
  * See [testing documentation](http://d.android.com/tools/testing).
  */
-@RunWith(AndroidJUnit4::class)
 class ExampleInstrumentedTest {
     @Test
     fun useAppContext() {
         // Context of the app under test.
-        val appContext = InstrumentationRegistry.getTargetContext()
+        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
         assertEquals("org.jlleitschuh.gradle.ktlint.android", appContext.packageName)
     }
 }