+++ samples/android-app/src/main/java/org/jlleitschuh/gradle/ktlint/android/ItemListActivity.kt
@@ -2,20 +2,15 @@ package org.jlleitschuh.gradle.ktlint.android
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.design.widget.Snackbar
-import android.support.v7.app.AppCompatActivity
-import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
-import kotlinx.android.synthetic.main.activity_item_list.fab
-import kotlinx.android.synthetic.main.activity_item_list.toolbar
-import kotlinx.android.synthetic.main.item_list.item_detail_container
-import kotlinx.android.synthetic.main.item_list.item_list
-import kotlinx.android.synthetic.main.item_list_content.view.content
-import kotlinx.android.synthetic.main.item_list_content.view.id_text
+import androidx.appcompat.app.AppCompatActivity
+import androidx.recyclerview.widget.RecyclerView
+import com.google.android.material.snackbar.Snackbar
 import org.jlleitschuh.gradle.ktlint.android.ItemListActivity.SimpleItemRecyclerViewAdapter.ViewHolder
+import org.jlleitschuh.gradle.ktlint.android.databinding.ActivityItemListBinding
 import org.jlleitschuh.gradle.ktlint.android.dummy.DummyContent
 
 /**
@@ -28,43 +23,30 @@ import org.jlleitschuh.gradle.ktlint.android.dummy.DummyContent
  */
 class ItemListActivity : AppCompatActivity() {
 
-    /**
-     * Whether or not the activity is in two-pane mode, i.e. running on a tablet
-     * device.
-     */
-    private var mTwoPane: Boolean = false
+    private var viewBinding: ActivityItemListBinding? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_item_list)
+        viewBinding = ActivityItemListBinding.inflate(LayoutInflater.from(this))
+        setContentView(viewBinding?.root)
 
-        setSupportActionBar(toolbar)
-        toolbar.title = title
+        setSupportActionBar(viewBinding?.toolbar)
+        viewBinding?.toolbar?.title = title
 
-        fab.setOnClickListener { view ->
+        viewBinding?.fab?.setOnClickListener { view ->
             Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
                 .setAction("Action", null).show()
         }
 
-        if (item_detail_container != null) {
-            // The detail container view will be present only in the
-            // large-screen layouts (res/values-w900dp).
-            // If this view is present, then the
-            // activity should be in two-pane mode.
-            mTwoPane = true
-        }
-
-        setupRecyclerView(item_list)
+        setupRecyclerView(viewBinding?.itemListContainer?.itemList)
     }
 
-    private fun setupRecyclerView(recyclerView: RecyclerView) {
-        recyclerView.adapter = SimpleItemRecyclerViewAdapter(this, DummyContent.ITEMS, mTwoPane)
+    private fun setupRecyclerView(recyclerView: RecyclerView?) {
+        recyclerView?.adapter = SimpleItemRecyclerViewAdapter(DummyContent.ITEMS)
     }
 
     class SimpleItemRecyclerViewAdapter(
-        private val mParentActivity: ItemListActivity,
         private val mValues: List<DummyContent.DummyItem>,
-        private val mTwoPane: Boolean
     ) : RecyclerView.Adapter<ViewHolder>() {
 
         private val mOnClickListener: View.OnClickListener
@@ -72,21 +54,10 @@ class ItemListActivity : AppCompatActivity() {
         init {
             mOnClickListener = View.OnClickListener { v ->
                 val item = v.tag as DummyContent.DummyItem
-                if (mTwoPane) {
-                    val fragment = ItemDetailFragment().apply {
-                        arguments = Bundle()
-                        arguments?.putString(ItemDetailFragment.ARG_ITEM_ID, item.id)
-                    }
-                    mParentActivity.supportFragmentManager
-                        .beginTransaction()
-                        .replace(R.id.item_detail_container, fragment)
-                        .commit()
-                } else {
-                    val intent = Intent(v.context, ItemDetailActivity::class.java).apply {
-                        putExtra(ItemDetailFragment.ARG_ITEM_ID, item.id)
-                    }
-                    v.context.startActivity(intent)
+                val intent = Intent(v.context, ItemDetailActivity::class.java).apply {
+                    putExtra(ItemDetailFragment.ARG_ITEM_ID, item.id)
                 }
+                v.context.startActivity(intent)
             }
         }
 
@@ -112,13 +83,8 @@ class ItemListActivity : AppCompatActivity() {
         }
 
         inner class ViewHolder(mView: View) : RecyclerView.ViewHolder(mView) {
-            val mIdView: TextView = mView.id_text
-            val mContentView: TextView = mView.content
-        }
-
-        override fun onViewRecycled(holder: ViewHolder?) {
-            System.err.println("OnViewRecycled: ${holder?.adapterPosition}")
-            super.onViewRecycled(holder)
+            val mIdView: TextView = mView.findViewById(R.id.id_text)
+            val mContentView: TextView = mView.findViewById(R.id.content)
         }
     }
 }